Renaming ERC20/ERC721 - tasit-action
