+++ packages/tasit-action/src/erc20/ERC20Full.js
@@ -1,7 +1,7 @@
 import Contract from "../contract/Contract";
-import abi from "../../../tasit-contracts/abi/ERC20Detailed.json";
+import abi from "../../../tasit-contracts/abi/MyERC20Full.json";
 
-export default class ERC20Detailed extends Contract {
+export default class ERC20Full extends Contract {
   constructor(address, wallet) {
     super(address, abi, wallet);
   }
