+++ packages/tasit-action/src/erc721/ERC721Full.js
@@ -1,5 +1,5 @@
 import Contract from "../contract/Contract";
-import abi from "../../../tasit-contracts/abi/ERC721Full.json";
+import abi from "../../../tasit-contracts/abi/MyERC721Full.json";
 
 export default class ERC721Full extends Contract {
   constructor(address, wallet) {
